requirements.txt
openai-whisper==20231117
pyaudio==0.2.14
pynput==1.7.6
numpy==1.24.3
torch==2.1.0
pydub==0.25.1
webrtcvad==2.0.10 